% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify2.R
\name{classify2}
\alias{classify2}
\title{do tensor clustering}
\usage{
classify2(x, k, r, l, lambda = 0, max.iter = 1000, threshold = 1e-15,
  trace = FALSE, Cs.init = NULL, Ds.init = NULL, Es.init = NULL,
  nstart = 20, method = "L0", center = FALSE)
}
\arguments{
\item{x}{a three-dimensional array}

\item{k}{the clusters number of mode 1}

\item{r}{the clusters number of mode 2}

\item{l}{the clusters number of mode 3}

\item{lambda}{a positive numeric value}

\item{max.iter}{maximum times of iteration}

\item{threshold}{...}

\item{trace}{...}

\item{Cs.init}{...}

\item{Ds.init}{...}

\item{Es.init}{...}

\item{nstart}{...}

\item{method}{two options: "L0", "L1". Two methods use different penalties, where "L1" indicating Lasso penalty.}

\item{center}{...}
}
\value{
a list \code{judgeX}
               \code{Cs} clustering result in mode 1 
               \code{Ds} clustering result in mode 2
               \code{Es} clustering result in mode 3
               \code{mus}
}
\description{
do tensor clustering
}
