% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.data.R
\name{get.data}
\alias{get.data}
\title{generate a random tensor}
\usage{
get.data(n, p, q, k, r, l, error = 3, sort = TRUE,
  sparse.percent = 0, center = FALSE, mumin = -3, mumax = 3)
}
\arguments{
\item{n}{the dimension of mode 1}

\item{p}{the dimension of mode 2}

\item{q}{the dimension of mode 3}

\item{k}{the clusters number of mode 1}

\item{r}{the clusters number of mode 2}

\item{l}{the clusters number of mode 3}

\item{error}{noise in rnorm}

\item{sort}{if TRUE, the data belongs to the same clusters would assemble together}

\item{sparse.percent}{the proportion of 0 mean}

\item{center}{if center sum(x) = 0}

\item{mumin}{...}

\item{mumax}{...}
}
\value{
a list \code{x}
               \code{truthX}
               \code{truthCs} true distribution in mode 1
               \code{truthDs} true distribution in mode 2
               \code{truthEs} true distribution in mode 3
               \code{mus} 
               \code{binaryX}
}
\description{
generate a random tensor
}
