% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation}
\alias{simulation}
\title{simulation: choose the best lambda when the range is given}
\usage{
simulation(n, p, q, k, r, l, error, lambda, iteration = 1,
  method = "L0")
}
\arguments{
\item{n}{the dimension of mode 1}

\item{p}{the dimension of mode 2}

\item{q}{the dimension of mode 3}

\item{k}{the clusters number of mode 1}

\item{r}{the clusters number of mode 2}

\item{l}{the clusters number of mode 3}

\item{error}{the standard deviation when producing data}

\item{lambda}{a vector of possible lambda, for example: lambda = c(0,50,100,200,300,400,500,600,700,800,900,1000,1100,1200)}

\item{iteration}{iteration times}

\item{method}{two options: "L0", "L1". Two methods use different penalties, where "L1" indicating Lasso penalty.}
}
\description{
choose the best lambda for a certain range.
}
