% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.evaluate.R
\name{sparse.evaluate}
\alias{sparse.evaluate}
\title{evaluate the accuracy of the model when inputting a sparse tensor}
\usage{
sparse.evaluate(bires, data)
}
\arguments{
\item{bires}{the return list of label2()}

\item{data}{the return list of get.data()}
}
\description{
evaluate the accuracy of the model when inputting a sparse tensor
}
