% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.calculateBIC.R
\name{tensor.calculateBIC}
\alias{tensor.calculateBIC}
\title{Calculate BIC}
\usage{
tensor.calculateBIC(x, clusterobj, method = "L0")
}
\arguments{
\item{x}{a three-dimensional array}

\item{clusterobj}{the return object of label2()}

\item{method}{two options: "L0", "L1". Two methods use different penalties, where "L1" indicating Lasso penalty.}
}
\value{
a vector [1]BIC, [2]nonzeromus
}
\description{
Calculate BIC
}
